backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
PyYAML==6.0.1
jsonschema==4.19.0